backend-infra/pyproject.toml
[project]
name = "backend-infra"
version = "0.1.0"
description = "YourClaw infrastructure management API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "kubernetes>=32.0",
    "hcloud>=2.16.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend_infra"]
